move messages controller to api namespace
